requirements.txt
netifaces
python-nmap
scapy